[quantized] Replace conv_p with convolution_op in qnnpack
convolution_op makes conv_param struct redundant, since it contains all the params of conv_param and more. We don't need to pass both structs to qnnpack or hold both in the packed weights, let's just hold convolution_op.

This makes it easier to implement 3dconv since we won't have to template two structs. The conv_param struct is left in existence since tests rely on it to set up the convolution.

Differential Revision: [D30738727](https://our.internmc.facebook.com/intern/diff/D30738727/)

ghstack-source-id: 139097792
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65783